Quote credentials (as per pymongo documentation)
When using a password tin the pymongo URI such that the password
contains special characters, authentication with MongoDB fails.

The pymongo documentation explains how to escape these special
characters (see
https://pymongo.readthedocs.io/en/stable/examples/authentication.html).

The password could technically be escaped directly in the URI on the
command-line, but this is very awkward and it would be quite convenient
if `pymongo-migrate` could do this automatically.